+++ src/plugins/weather_station/moon_phase.h
@@ -62,6 +62,9 @@ struct MoonPhaseType
         else if (iPhase > 0.5) {
             cpPh -= (0.25 / 8.0);
         }
+        else if (iPhase < 0.125) {
+            cpPh += (0.5 / 8.0);
+        }
         else if (iPhase < 0.25) {
             cpPh += (0.25 / 8.0);
         }